% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatar_valor_tabela.R
\name{formatar_valor_individual}
\alias{formatar_valor_individual}
\title{Formata valores monetários com escala dinâmica}
\usage{
formatar_valor_individual(x, accuracy = 0.01)
}
\arguments{
\item{x}{Vetor numérico contendo os valores a serem formatados.}

\item{accuracy}{Precisão da formatação. Default: 0.1.}
}
\value{
Um vetor de caracteres com os valores formatados.
}
\description{
Esta função formata um vetor numérico de valores monetários, ajustando a escala e o sufixo
de acordo com o valor máximo do vetor. Se o valor máximo for maior ou igual a 1 bilhão,
utiliza a escala de 1e-9 e o sufixo "B"; se for maior ou igual a 1 milhão, usa 1e-6 e "M";
se for maior ou igual a 1 mil, usa 1e-3 e "K"; caso contrário, não aplica escala.
}
\examples{
formatar_valor_individual(c(500, 200000, 3000000000))

}
