% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatar_valor_tabela.R
\name{formatar_valor_tabela_produto}
\alias{formatar_valor_tabela_produto}
\title{Formata uma coluna inteira de valores monetários}
\usage{
formatar_valor_tabela_produto(x, accuracy = 0.01)
}
\arguments{
\item{x}{Vetor numérico contendo os valores a serem formatados.}

\item{accuracy}{Precisão da formatação. Default: 0.1.}
}
\value{
Um vetor de caracteres com os valores formatados.
}
\description{
Esta função aplica `formatar_valor_individual` a uma coluna inteira de valores monetários.
}
\examples{
formatar_valor_tabela_produto(c(500, 200000, 3000000000))

}
