% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatorio_brasil_pais_anual.R
\name{relatorio_brasil_pais_anual}
\alias{relatorio_brasil_pais_anual}
\title{Relatório de relações bilaterais comerciais Brasil-país}
\usage{
relatorio_brasil_pais_anual(pais)
}
\description{
A função gera um relatório sobre os dados de comércio do Brasil com o país especificado somente com a parte anual
}
